+++ src/components/Button/ButtonTypes.tsx
@@ -1,6 +1,6 @@
 export enum ButtonTypes {
-    Outline = "outline",
-    Filled = "filled",
+    Primary = "filled",
+    Secondary = "outline",
     Link = "link",
 }