+++ src/components/Card/Card.test.tsx
@@ -33,7 +33,7 @@ describe("Card", () => {
                         console.log(this);
                     }}
                     id="button1"
-                    buttonType={ButtonTypes.Filled}
+                    buttonType={ButtonTypes.Primary}
                     type="submit"
                 >
                     Example CTA