+++ src/components/SearchResultItem/SearchResultItem.test.tsx
@@ -49,7 +49,7 @@ describe("SearchResultItem", () => {
                                     console.log(this);
                                 }}
                                 id="button1"
-                                buttonType={ButtonTypes.Filled}
+                                buttonType={ButtonTypes.Primary}
                                 type="submit"
                             >
                                 Example CTA